refactor joystick and button components
