+++ asteroids/lib/src/components/components.dart
@@ -1,5 +1,5 @@
 export 'asteroid.dart';
 export 'player.dart';
 export 'shot.dart';
-export 'virtual_joystick.dart';
-export 'virtual_button.dart';
+export 'joystick.dart';
+export 'game_button.dart';
