+++ asteroids/lib/src/components/player.dart
@@ -1,6 +1,5 @@
 import 'package:flame/collisions.dart';
 import 'package:flame/components.dart';
-import 'package:flame/effects.dart';
 import 'package:flame/extensions.dart';
 
 import 'package:flutter/material.dart';
